Remove unneeded index lines from patches
Removes all index lines for text patches.

New method:
$ sed -i -z -E 's/\nindex [a-f0-9]{40}\.\.[a-f0-9]{40}( [0-9]+)?\n---/\n---/g' *.patch

We already tried to do this, but it didn't handle index lines for new
files. We can't remove them all, either, because binary patches need
them. It seems to be safe to remove any index lines that precede a text
patch, whose first line always begins with ---, so we do that. This may
require GNU sed due to the use of the -z option which treats the entire
file as one line so that we can easily operate on multiple lines.

Change-Id: I889a69d654722bf0d897bdb088fa308e64782212