Fix crash on first startup in EEA
Solves an issue in which, with a system language whose country code is
a country within the European Economic Area (extension of the EU), the
browser may crash in template_url_prepopulate_data.cc with this error:
`Check failed: t_urls.size() <= kMaxEeaPrepopulatedEngines (10 vs. 8)`

Discovered by a user in the CalyxOS Development channel. Does not appear
to affect Cromite itself for some reason, but may impact forks.

Issue: calyxos#2636
Change-Id: Ic9363a247963b47edff428ebef2c8c4026ddc7b5