Build environment script and instructions
Includes helper functions and state tracking for the building
and porting process. Major version porting procedures are intended
for interactive use.

Also adds a README.md with some instructions on how to use this.

Both the README.md and the script are a work in progress and may be
updated over time.

Issue: calyxos#3141
Issue: calyxos#1719
Change-Id: I4fadb2d9c5fae58f49c317917406c96f7e8565fd